// src/app/models/ques-Ans-model.ts
export class QuesAnsModel{
    id: number | any;
    levelId: number | any;
    questionType: string | any;
    serviceId: string | any;
    portalId: string | any;
    quizDuration: number | any;
    question: string | any;
    option1: string  | any;
    option2: string | any;
    option3: string | any;
    option4: string | any;
    correctAnswer: string | any ;
    trivia: string | any;
    ststus: boolean  | any;
    triviaAr: string  | undefined
triviaFr:  string  | undefined;
questionAr:  string  | undefined
questionFr:  string  | undefined;
option4Ar:  string  | undefined
option4Fr:  string  | undefined;
option3Ar:  string  | undefined
option3Fr:  string  | undefined;
option2Ar:  string  | undefined
option2Fr:  string  | undefined;
option1Ar:  string  | undefined
option1Fr:  string  | undefined
levelCoins: number | string | undefined
}